--- test-toolkit/docker/1.4.0/Dockerfile.dlc.gpu
@@ -1,10 +1,0 @@
-ARG region
-from 763104351884.dkr.ecr.$region.amazonaws.com/pytorch-training:1.4.0-gpu-py3
-
-COPY docker/build_artifacts/changehostname.c /
-COPY docker/build_artifacts/start_with_right_hostname.sh /usr/local/bin/start_with_right_hostname.sh
-RUN chmod +x /usr/local/bin/start_with_right_hostname.sh
-
-COPY dist/sagemaker_pytorch_training-*.tar.gz /sagemaker_pytorch_training.tar.gz
-RUN pip install --upgrade --no-cache-dir /sagemaker_pytorch_training.tar.gz && \
-    rm /sagemaker_pytorch_training.tar.gz
