--- test-toolkit/docker/1.4.0/Dockerfile.pytorch
@@ -1,20 +1,0 @@
-from pytorch/pytorch:1.4-cuda10.1-cudnn7-runtime
-
-RUN apt-get update \
- && apt-get install -y --no-install-recommends jq \
- && rm -rf /var/lib/apt/lists/*
-
-COPY docker/build_artifacts/changehostname.c /
-COPY docker/build_artifacts/start_with_right_hostname.sh /usr/local/bin/start_with_right_hostname.sh
-RUN chmod +x /usr/local/bin/start_with_right_hostname.sh
-
-COPY dist/sagemaker_pytorch_training-*.tar.gz /sagemaker_pytorch_training.tar.gz
-RUN pip install --no-cache-dir /sagemaker_pytorch_training.tar.gz && \
-    rm /sagemaker_pytorch_training.tar.gz
-
-ENV SAGEMAKER_TRAINING_MODULE=sagemaker_pytorch_container.training:main
-
-WORKDIR /
-
-# Starts framework
-ENTRYPOINT ["bash", "-m", "start_with_right_hostname.sh"]
