+++ test/integration/local/test_single_machine_training.py
@@ -14,17 +14,16 @@ from __future__ import absolute_import
 
 import os
 
-import pytest
 from sagemaker.pytorch import PyTorch
 
-from test.utils.local_mode_utils import assert_files_exist
-from test.integration import data_dir, fastai_path, fastai_mnist_script, mnist_script, PYTHON3, ROLE
+from utils.local_mode_utils import assert_files_exist
+from integration import data_dir, mnist_script, ROLE
 
 
-def test_mnist(docker_image, processor, instance_type, sagemaker_local_session, tmpdir):
+def test_mnist(image_uri, processor, instance_type, sagemaker_local_session, tmpdir):
     estimator = PyTorch(entry_point=mnist_script,
                         role=ROLE,
-                        image_name=docker_image,
+                        image_name=image_uri,
                         train_instance_count=1,
                         train_instance_type=instance_type,
                         sagemaker_session=sagemaker_local_session,
@@ -32,22 +31,6 @@ def test_mnist(docker_image, processor, instance_type, sagemaker_local_session, 
                         output_path='file://{}'.format(tmpdir))
 
     _train_and_assert_success(estimator, data_dir, str(tmpdir))
-
-
-def test_fastai_mnist(docker_image, instance_type, py_version, sagemaker_local_session, tmpdir):
-    if py_version != PYTHON3:
-        pytest.skip('Skipping the test because fastai supports >= Python 3.6.')
-
-    estimator = PyTorch(entry_point=fastai_mnist_script,
-                        role=ROLE,
-                        image_name=docker_image,
-                        train_instance_count=1,
-                        train_instance_type=instance_type,
-                        sagemaker_session=sagemaker_local_session,
-                        output_path='file://{}'.format(tmpdir))
-
-    input_dir = os.path.join(fastai_path, 'mnist_tiny')
-    _train_and_assert_success(estimator, input_dir, str(tmpdir))
 
 
 def _train_and_assert_success(estimator, input_dir, output_path):
